Point_Raster_Locator
Takes a DEM, a set of rasters, and set of points. It uses the elevation at each point (from DEM) to calculate the which raster the point is closest to. It was developed to find which aquifer a well is drawing from, and such will find the closest raster above the point. But this can be changed to nearest or closest below.